+++ myhome-shutter.js
@@ -27,7 +27,7 @@ module.exports = function (RED) {
 
     //////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
     // Add listener on node linked to a dedicated function call to be able to remove it on close
-    const listenerFunction = function (frame) {
+    const listenerFunction = function (ownFamilyName , frame) {
       let msg = {};
       node.processReceivedBUSFrames (msg, frame);
     };
@@ -91,6 +91,19 @@ module.exports = function (RED) {
           if (!Array.isArray(frame)) {
             payload.command_received = frame;
           }
+          // MSG1 : add major node configuration info on both returned message
+          msg.mh_nodeConfigInfo = {
+          	'name' : config.name ,
+          	'topic' : config.topic ,
+          	'buslevel' : config.buslevel ,
+          	'shutterid' : config.shutterid ,
+          	'isgroup' : config.isgroup ,
+          	'gateway' : {
+          		'name' : gateway.name ,
+          		'host' : gateway.host ,
+          		'port' : gateway.port
+          	}
+          };
           // MSG1 : Add all current node stored values to payload
           payload.state = payloadInfo.state;
           // MSG1 : Add misc info